refactor: application test 통과를 위한 app 분리
