+++ src/index.js
@@ -1,9 +1,9 @@
 import App from './App.js';
-import Calculator from './Calculator.js';
-import Delimiter from './Delimiter.js';
+import Calculator from './app/Calculator.js';
+import Delimiter from './app/Delimiter.js';
 import SchemaValidator from './lib/SchemaValidator.js';
-import Validator from './Validator.js';
-import View from './View.js';
+import Validator from './app/Validator.js';
+import View from './app/View.js';
 
 const app = new App(
   new View(),
